Méga nettoyage pour être synchro avec ta version
A partir de là, j'essayerai de faire des petites branches par fonctionnalité (je dis bien : essayerai :D)
