+++ NWScript/Types/nwcreature.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using NWN.Enums;
